requirements.txt
binance-futures-connector
configparser